Update modules with new methods
