+++ src/shipchain_common/exceptions.py
@@ -188,6 +188,19 @@ def invalid_url(request, exception, *args, **kwargs):
     return JsonResponse(data, status=status.HTTP_404_NOT_FOUND)
 
 
+class AWSIoTError(APIException):
+    status_code = status.HTTP_500_INTERNAL_SERVER_ERROR
+    default_detail = 'Internal Service Error.'
+    default_code = 'server_error'
+
+    def __init__(self, detail, status_code=None, code=None):
+        super(AWSIoTError, self).__init__(detail, code)
+        self.detail = detail
+
+        if status_code:
+            self.status_code = status_code
+
+
 class RPCError(APIException):
     status_code = status.HTTP_500_INTERNAL_SERVER_ERROR
     default_detail = 'Internal Service Error.'
